Travis: Refactor out ds3_lock_twice
Large methods need to be in separate modules.
